integrate xbase expressions in execution language
This commit removes the ActionInstance and ParameterValue classes, and replace them with XExpressions. This gives more flexibility when defining a bot: XExpressions can be used within execution rules to benefit from the expressivity of xbase as well as auto-completion, syntax checking, and parameter types checks. This update changes a bit how action are invoked: the name of the parameters are not required anymore when calling an action, as an example:
`SlackPlatform.Reply(message : "myMessage")`
is now called with the following syntax:
`SlackPlatform.Reply("myMessage")`