moved System.out.println to log4j (fix #15)
Xtext already uses log4j 1, so it was easier to keep the same logger for the eclipse-related code. The logger is set to level INFO when the classes are loaded in an eclipse context. The runtime component contains a log4j.properties file that disable eclipse-related logging when running a bot.